requirements.txt
Flask==3.0.2
boto3==1.34.122
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
